Refactor topology
